Revert "xwm: do not include shadow in input region"
This reverts commit 1c8ad8996a5523cf5af7dfd7ec61c83da3c13adf.
And re-introduces the bug it was intended to fix, see:
https://lists.freedesktop.org/archives/wayland-devel/2017-December/036402.html

Reverting this because it causes harm to all xwayland clients - the
input region no longer gets adjusted when resizing windows.

start an xterm, resize it larger, you can no longer interact with the
new area of the window (including the server side decor).