contributing: add copy of DCO
The DCO was fetched from https://developercertificate.org/ on Feb 4, 2019.
DCO-1.1.txt is a verbatim copy of it.

We should carry a copy of this legal document to be explicit on what wording we
refer to. A link to the URL is not enough, because it may go stale or change
contents without notice.

This clarifies on what the S-o-b actually means. The meaning itself is the same
as before.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>